allan/exemplo_3/flask_app/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
crewai>=0.28.8
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
markdown>=3.5.1
pdfkit>=1.0.0